requirements.txt
streamlit
ultralytics
Pillow
torch
opencv-python